+++ order-microservice/src/main/java/com/metao/book/order/application/cart/UpdateCartItemQtyDTO.java
@@ -1,6 +1,4 @@
 package com.metao.book.order.application.cart;
 
-import java.math.BigDecimal;
-
-public record UpdateCartItemQtyDTO(BigDecimal quantity) {
+public record UpdateCartItemQtyDTO(int quantity) {
 }
